9: Refactored Intcode computer a bit, hitting 203 error now

+++ 2019/IntCode Computer/Day7/Program.cs
@@ -52,7 +52,7 @@ namespace Day7
 							amps[i].Run();
 							if (amps[i].HasOutput)
 							{
-								nextInput = amps[i].GetOutput();
+								nextInput = (int)amps[i].GetOutput();
 								Console.WriteLine($"Retreieved {nextInput} from Amp {ampNames[i]}");
 							}
 						}
